WIP: Still need to work out kinks in generator correlation
