--- skxray/core/correlation/__init__.py
@@ -1,1 +1,0 @@
-from .correlation import multi_tau_auto_corr, auto_corr_scat_factor
